feat: 애노테이션 기반 인증 개선
#18
