+++ src/main/java/atwoz/atwoz/common/auth/context/AuthResolver.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.common.auth;
+package atwoz.atwoz.common.auth.context;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
