--- src/main/java/atwoz/atwoz/common/auth/exception/TokenNotExistException.java
@@ -1,7 +1,0 @@
-package atwoz.atwoz.common.auth.exception;
-
-public class TokenNotExistException extends RuntimeException {
-    public TokenNotExistException() {
-        super("토큰이 존재하지 않습니다.");
-    }
-}
